# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/habits": {
      "get": {
        "operationId": "HabitsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "HabitsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/v1/habits/{id}/tick": {
      "post": {
        "operationId": "HabitsController_tick",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/v1/habits/{id}": {
      "patch": {
        "operationId": "HabitsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "HabitsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/tasks": {
      "get": {
        "operationId": "TasksController_list",
        "summary": "List user tasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tasks retrieved"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "TasksController_create",
        "summary": "Create new task",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Task created"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/tasks/{id}/complete": {
      "post": {
        "operationId": "TasksController_complete",
        "summary": "Mark task as completed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task completed"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/tasks/{id}": {
      "delete": {
        "operationId": "TasksController_delete",
        "summary": "Delete task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/alarms": {
      "get": {
        "operationId": "AlarmsController_list",
        "summary": "List user alarms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Alarms retrieved"
          }
        },
        "tags": [
          "Alarms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AlarmsController_create",
        "summary": "Create new alarm",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Alarm created"
          }
        },
        "tags": [
          "Alarms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/alarms/{id}": {
      "delete": {
        "operationId": "AlarmsController_delete",
        "summary": "Delete alarm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alarm deleted"
          }
        },
        "tags": [
          "Alarms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/alarms/{id}/dismiss": {
      "post": {
        "operationId": "AlarmsController_dismiss",
        "summary": "Dismiss alarm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alarm dismissed"
          }
        },
        "tags": [
          "Alarms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/alarms/{id}/fire": {
      "post": {
        "operationId": "AlarmsController_fire",
        "summary": "Fire alarm with voice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alarm fired with voice"
          }
        },
        "tags": [
          "Alarms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/voice/tts": {
      "post": {
        "operationId": "VoiceController_generateTTS",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/v1/brief": {
      "get": {
        "operationId": "BriefController_getBrief",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/brief/today": {
      "get": {
        "operationId": "BriefController_getTodaysBrief",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/nudge": {
      "get": {
        "operationId": "NudgesController_getNudge",
        "summary": "Get AI nudge based on user progress",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Nudge retrieved"
          }
        },
        "tags": [
          "Nudges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/chat": {
      "post": {
        "operationId": "NudgesController_sendChatMessage",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Nudges"
        ]
      }
    },
    "/v1/reports/daily": {
      "get": {
        "operationId": "ReportsController_getDailyReport",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/reports/weekly": {
      "get": {
        "operationId": "ReportsController_getWeeklyReport",
        "parameters": [
          {
            "name": "week",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/reports/morning-brief": {
      "get": {
        "operationId": "ReportsController_getMorningBrief",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/reports/evening-summary": {
      "get": {
        "operationId": "ReportsController_getEveningSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "DrillSergeant API",
    "description": "Production backend for DrillSergeant mobile app",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "IdempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
    "schemas": {}
  }
}